+++ src/objects.js
@@ -1385,13 +1385,15 @@ SpriteMorph.prototype.initBlocks = function () {
             spec: 'numbers from %n to %n',
             defaults: [1, 10]
         },
-        reportListCombination: {
+    /*
+        reportListCombination: { // currently not in use
             type: 'reporter',
             category: 'lists',
             spec: '%mlfunc %lists',
             defaults: [['append']]
         },
-        reportConcatenatedLists: { // deprecated
+    */
+        reportConcatenatedLists: {
             type: 'reporter',
             category: 'lists',
             spec: 'append %lists'
@@ -1606,11 +1608,6 @@ SpriteMorph.prototype.initBlockMigrations = function () {
         reportListLength: {
             selector: 'reportListAttribute',
             inputs: [['length']],
-            offset: 1
-        },
-        reportConcatenatedLists: {
-            selector: 'reportListCombination',
-            inputs: [['append']],
             offset: 1
         }
     };
@@ -2794,7 +2791,7 @@ SpriteMorph.prototype.blockTemplates = function (category) {
         blocks.push('-');
         blocks.push(block('doForEach'));
         blocks.push('-');
-        blocks.push(block('reportListCombination'));
+        blocks.push(block('reportConcatenatedLists'));
         blocks.push(block('reportReshape'));
         blocks.push('-');
         blocks.push(block('doAddToList'));
@@ -2971,7 +2968,7 @@ SpriteMorph.prototype.freshPalette = function (category) {
                         'reportCDR',
                         'reportListAttribute',
                         'reportListIndex',
-                        'reportListCombination',
+                        'reportConcatenatedLists',
                         'reportReshape',
                         'reportListContainsItem',
                         'reportListIsEmpty',
@@ -8955,7 +8952,7 @@ StageMorph.prototype.blockTemplates = function (category) {
         blocks.push('-');
         blocks.push(block('doForEach'));
         blocks.push('-');
-        blocks.push(block('reportListCombination'));
+        blocks.push(block('reportConcatenatedLists'));
         blocks.push(block('reportReshape'));
         blocks.push('-');
         blocks.push(block('doAddToList'));