+++ HISTORY.md
@@ -17,6 +17,7 @@
 * threads: enhanced list attributes 'rank', 'shape' and 'ravel' to also handle scalars
 * threads: enhanced 'reshape' to also handle scalars
 * lists: limit crash-dangerous matrix-exploding ops to 1 MM elements (reshape, crossproduct)
+* objects, threads: took out "crossproduct" primitive option from the palette
 
 ### 2021-02-08
 * lists, objects, threads: new RESHAPE primitive